requirements.txt
flask
gunicorn
mysql-connector-python
torch>=2.0
numpy